refactor: standardize on 'sign out' terminology
