+++ dev-client/src/components/modals/SignOutModal.tsx
@@ -15,24 +15,31 @@
  * along with this program. If not, see https://www.gnu.org/licenses/.
  */
 
+import {useCallback} from 'react';
 import {useTranslation} from 'react-i18next';
 
-import {ListButton} from 'terraso-mobile-client/components/buttons/list/ListButton';
-import {LogoutModal} from 'terraso-mobile-client/components/modals/LogoutModal';
+import {signOut} from 'terraso-client-shared/account/accountSlice';
 
-export function LogOutButton() {
+import {ConfirmModal} from 'terraso-mobile-client/components/modals/ConfirmModal';
+import {ModalProps} from 'terraso-mobile-client/components/modals/Modal';
+import {useDispatch} from 'terraso-mobile-client/store';
+
+export type LogoutModalProps = Pick<ModalProps, 'trigger'>;
+
+export function SignOutModal({trigger}: LogoutModalProps) {
   const {t} = useTranslation();
+  const dispatch = useDispatch();
+  const onSignOut = useCallback(() => {
+    dispatch(signOut());
+  }, [dispatch]);
 
   return (
-    <LogoutModal
-      trigger={onOpen => (
-        <ListButton
-          type="default"
-          iconName="logout"
-          labelText={t('settings.log_out')}
-          onPress={onOpen}
-        />
-      )}
+    <ConfirmModal
+      trigger={trigger}
+      body={t('sign_out.confirm_body')}
+      actionName={t('sign_out.confirm_action')}
+      isConfirmError={false}
+      handleConfirm={onSignOut}
     />
   );
 }
